Refactor Hudi BaseConnectorTests
The base test classes shouldn't provide the QueryRunner to make sure
that subclasses configure things correctly and the base class is more
re-usable by just providing the TestingConnectorBehavior and any test
overrides as needed.

This change also allows to remove BaseHudiMinioConnectorTest which is
now redundant.
