+++ plugin/trino-hudi/src/test/java/io/trino/plugin/hudi/BaseHudiConnectorTest.java
@@ -13,34 +13,20 @@
  */
 package io.trino.plugin.hudi;
 
-import com.google.common.collect.ImmutableMap;
 import io.trino.plugin.hudi.testing.TpchHudiTablesInitializer;
 import io.trino.testing.BaseConnectorTest;
-import io.trino.testing.QueryRunner;
 import io.trino.testing.TestingConnectorBehavior;
-import org.apache.hudi.common.model.HoodieTableType;
 import org.testng.annotations.Test;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static io.trino.plugin.hudi.HudiQueryRunner.createHudiQueryRunner;
 import static org.apache.hudi.common.model.HoodieRecord.HOODIE_META_COLUMNS;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public abstract class BaseHudiConnectorTest
         extends BaseConnectorTest
 {
-    @Override
-    protected QueryRunner createQueryRunner()
-            throws Exception
-    {
-        return createHudiQueryRunner(
-                ImmutableMap.of(),
-                ImmutableMap.of("hudi.columns-to-hide", columnsToHide()),
-                new TpchHudiTablesInitializer(getHoodieTableType(), REQUIRED_TPCH_TABLES));
-    }
-
     @SuppressWarnings("DuplicateBranchesInSwitch")
     @Override
     protected boolean hasBehavior(TestingConnectorBehavior connectorBehavior)
@@ -93,9 +79,7 @@ public abstract class BaseHudiConnectorTest
                         ")");
     }
 
-    protected abstract HoodieTableType getHoodieTableType();
-
-    static String columnsToHide()
+    protected static String columnsToHide()
     {
         List<String> columns = new ArrayList<>(HOODIE_META_COLUMNS.size() + 1);
         columns.addAll(HOODIE_META_COLUMNS);
