+++ plugin/trino-hudi/src/test/java/io/trino/plugin/hudi/TestHudiCopyOnWriteConnectorTest.java
@@ -13,16 +13,23 @@
  */
 package io.trino.plugin.hudi;
 
-import org.apache.hudi.common.model.HoodieTableType;
+import com.google.common.collect.ImmutableMap;
+import io.trino.plugin.hudi.testing.TpchHudiTablesInitializer;
+import io.trino.testing.QueryRunner;
 
+import static io.trino.plugin.hudi.HudiQueryRunner.createHudiQueryRunner;
 import static org.apache.hudi.common.model.HoodieTableType.COPY_ON_WRITE;
 
 public class TestHudiCopyOnWriteConnectorTest
         extends BaseHudiConnectorTest
 {
     @Override
-    protected HoodieTableType getHoodieTableType()
+    protected QueryRunner createQueryRunner()
+            throws Exception
     {
-        return COPY_ON_WRITE;
+        return createHudiQueryRunner(
+                ImmutableMap.of(),
+                ImmutableMap.of("hudi.columns-to-hide", columnsToHide()),
+                new TpchHudiTablesInitializer(COPY_ON_WRITE, REQUIRED_TPCH_TABLES));
     }
 }
